+++ FlowDesigner/data-flow/src/Add.cc
@@ -58,7 +58,7 @@ public:
    {
       ObjectRef ReturnValue = getInput(m_input1ID, count);
 
-      for (int j = 1; j < inputs.size(); j++)
+      for (size_t j = 1; j < inputs.size(); j++)
       {
          ObjectRef inputValue = getInput(j, count);
          ReturnValue = ReturnValue + inputValue;
