+++ FlowDesigner/data-flow/src/OR.cc
@@ -72,7 +72,7 @@ public:
    {
       bool outValue = false;
 
-      for (int j = 0; j < inputs.size(); j++)
+      for (size_t j = 0; j < inputs.size(); j++)
       {
          ObjectRef inputValue = getInput(j, count);
          if (dereference_cast<bool> (inputValue) && !outValue)
